teacher/16/shop/main/translation.py
from modeltranslation.translator import translator, TranslationOptions
from .models import Product

class ProductTranslationOptions(TranslationOptions):
    fields = ('name', 'desc')

translator.register(Product, ProductTranslationOptions)