teacher/16/requirements.txt
Django==3.2.8
Pillow==8.4.0
django-image-cropping==1.6.1
easy-thumbnails==2.7.2
django-tinymce==3.3.0
django-rosetta
django-modeltranslation